Restructure component folder.
